refactor: centralize utilities and improve docs
