+++ src/utils.ts
@@ -1,0 +1,34 @@
+export interface Sector {
+  label: string;
+  icon: string;
+  angle: number;
+}
+
+// Convert polar coordinates to cartesian coordinates for SVG positioning
+export function polarToCartesian(
+  cx: number,
+  cy: number,
+  r: number,
+  angleInDegrees: number
+) {
+  const angleInRadians = ((angleInDegrees - 90) * Math.PI) / 180;
+  return {
+    x: cx + r * Math.cos(angleInRadians),
+    y: cy + r * Math.sin(angleInRadians),
+  };
+}
+
+// Generate an SVG arc path string for a circular segment
+export function describeArc(
+  cx: number,
+  cy: number,
+  r: number,
+  startAngle: number,
+  endAngle: number
+) {
+  const start = polarToCartesian(cx, cy, r, endAngle);
+  const end = polarToCartesian(cx, cy, r, startAngle);
+  const largeArcFlag = endAngle - startAngle <= 180 ? "0" : "1";
+  return `M ${start.x} ${start.y} A ${r} ${r} 0 ${largeArcFlag} 0 ${end.x} ${end.y}`;
+}
+
